+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Version 7.0.0-dev
 - BC break: Tester's constructor now takes TestsFolderProvider as its first parameter
 - BC break: removed Tester::$useColors
 - BC break: removed (virtual) method Job::onAfterExecute
+- BC break: removed BaseAttribute
 
 Version 6.1.0
 - added CodeCoverageExtension, Reader, IAnnotationsReaderEngine, PhpAttributesEngine, TAssertions::getCounter(), TestsStartedEvent, TestsFinishedEvent to public api